requirements.txt
langchain
langchain-huggingface
langchain-groq
huggingface_hub
python-dotenv
streamlit
cassio
datasets
sentence-transformers
PyPDF2
